requirements.txt
torch>=2.1.0
torchvision>=0.16.0
albumentations>=1.4.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0
pyyaml>=6.0.1
matplotlib>=3.8.0
tqdm>=4.66.0
kaggle>=1.6.0
